changes_to_gads/requirements.txt
Flask==2.2.2
gunicorn==20.1.0
pandas==1.5.3
google-ads==19.0.0
google-cloud-bigquery==3.3.0
google-cloud-secret-manager==2.12.0